MODIFY: Put and Read from error channel first
Reading from the information channel can cause deadlocks.

In Go, channels block until both a sender and receiver are ready. This means that if you try to read from the `shortUrl` channel before the `errorChannel`, and an error occurs, your program will deadlock.

Here's why:

1. If the URL is invalid, the function will send an error to the `errorChannel` and then exit without ever sending anything to the `shortUrl` channel.
2. If you try to read from the `shortUrl` channel first, your program will block because nothing will ever be sent to that channel.
3. Meanwhile, the error is waiting in the `errorChannel`, but it can't be received because your program is blocked waiting for `shortUrl`.

This situation is a deadlock, and Go will panic when this happens.

To avoid this, you should always read from the `errorChannel` first. If an error occurred (i.e., the received error is not `nil`), you can return early and avoid reading from the `shortUrl` channel altogether. If no error occurred (i.e., the received error is `nil`), it's safe to read from the `shortUrl` channel because you know that something will definitely be sent to it.